backend/requirements.txt
ultralytics
opencv-python
flask
flask-cors
requests
python-dotenv
numpy